Separada la logida de JavaScript en un hook personalizado: useHeaderLogic.js
